refactor: save to same file
